Move common utility functions to util.py file.
