+++ podocytes/main.py
@@ -21,6 +21,8 @@ from skimage.feature import blob_dog
 
 import tifffile._tifffile  # imported to silence pims warning
 
+from util import find_files
+
 
 def main():
     args = configure_parser()  # User input arguments
@@ -212,31 +214,6 @@ def filter_by_size(label_image, min_diameter, max_diameter):
                 (region.equivalent_diameter <= max_diameter)):
             regions.append(region)
     return regions
-
-
-def find_files(input_directory, ext):
-    """Recursive search for filenames matching specified extension.
-
-    Parameters
-    ----------
-    input_directory : str
-        Filepath to input directory location.
-    ext : str
-        File extension (eg: '.tif', '.lif', etc.)
-
-    Returns
-    -------
-    filelist : list of str
-        List of files matching specified extension in the input directory path.
-    """
-    filelist = []
-    for root, _, files in os.walk(input_directory):
-        for f in files:
-            # ignore hidden files
-            if f.endswith(ext) and not f.startswith('.'):
-                filename = os.path.join(root, f)
-                filelist.append(filename)
-    return filelist
 
 
 def find_podocytes(podocyte_image, glomeruli_region,
